requirements.txt
click==8.1.7
Flask==3.0.3
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
tensorflow==2.16.1
typing-extensions==4.7.1
Werkzeug==3.0.0
zipp==3.15.0
Pillow==9.0.1
